pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["project"]

[project]
name = "hammer_systems"
version = "0.4.0"
description = "Тестоввое задание для компании Hammer Systems."
requires-python = ">=3.13"
dependencies = [
    "djangorestframework>=3.16.0",
    "python-dotenv>=1.1.1",
    "psycopg==3.2.9",
    "psycopg-binary==3.2.9",
    "redis==6.2.0",
]

[project.optional-dependencies]
dev = ["ruff", "pre-commit", "pipdeptree"]

[tool.ruff.lint]
select = ["E", "F", "DJ", "I", "B"]
exclude = ["venv", ".venv", "__pycache__", "**/migrations/**"]

[tool.ruff.lint.pycodestyle]
max-line-length = 79
